Function ruma_signatures::content_hash

pub fn content_hash(
    object: &CanonicalJsonObject
) -> Result<Base64<Standard, [u8; 32]>, Error>

Creates a content hash for an event.

The content hash of an event covers the complete event including the unredacted contents. It is used during federation and is described in the Matrix server-server specification.

Parameters

object: A JSON object to generate a content hash for.

Errors

Returns an error if the event is too large.